pyproject.toml
[tool.poetry]
name = "ragger"
version = "0.1.0"
description = ""
authors = ["dewmal <dewmal@syigen.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ceylon = "^0.15.11"
fastapi = "^0.112.0"
python-socketio = "^5.11.3"
uvicorn = "^0.30.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
